WPr: pass JobContext DTO to job handler
So far the WorkProcessor provided the job's source queue name
as second argument for the job handler callback.

Passing a DTO object that contains this source queue name
but may contain any number of additional getters
is more practical
as it can be extended in the future
without breaking anything.

This breaks backwards compatibility
for all users which use the second argument in their job handlers,
but this is still v0 so it's not too bad.